Broke up index.js into modules
